+++ src/process.js
@@ -3,66 +3,65 @@ var messages = require("./messages.js");
 
 exports.removePlugins = function(config, options) {
     if (options.noRemove) {
-        console.log(messages.consoleMessages.skipRemove);
+        messages.outputMessage(messages.consoleMessages.skipRemove);
     } else {
-        console.log(messages.consoleMessages.findingInstalled);
+        messages.outputMessage(messages.consoleMessages.findingInstalled);
         var iplugins = methods.findInstalledPlugins();
 
-        console.log(messages.consoleMessages.removingInstalled);
+        messages.outputMessage(messages.consoleMessages.removingInstalled);
         methods.removePlugins(iplugins, options);
     }
 };
 
 exports.addPlugins = function(config, options) {
     if (options.noAdd) {
-        console.log(messages.consoleMessages.skipAdd);
+        messages.outputMessage(messages.consoleMessages.skipAdd);
     } else {
-        console.log(messages.consoleMessages.findingConfigPlugins);
+        messages.outputMessage(messages.consoleMessages.findingConfigPlugins);
         var plugins = methods.findConfigPlugins(config);
 
-        console.log(messages.consoleMessages.installingConfigPlugins);
+        messages.outputMessage(messages.consoleMessages.installingConfigPlugins);
         methods.installPlugins(plugins, options);
     }
 };
 
 exports.removePlatforms = function(config, options) {
     if (options.noRemove) {
-        console.log(messages.consoleMessages.skipRemove);
+        messages.outputMessage(messages.consoleMessages.skipRemove);
     } else {
-        console.log(messages.consoleMessages.findInstalledPlatforms);
+        messages.outputMessage(messages.consoleMessages.findInstalledPlatforms);
         var iplatforms = methods.findInstalledPlatforms();
         
-        console.log(messages.consoleMessages.removingInstalledPlatforms);
+        messages.outputMessage(messages.consoleMessages.removingInstalledPlatforms);
         methods.removePlatforms(iplatforms, options);
     }
 };
 
 exports.addPlatforms = function(config, options) {
     if (options.noAdd) {
-        console.log(messages.consoleMessages.skipAdd);
+        messages.outputMessage(messages.consoleMessages.skipAdd);
     } else {
-        console.log(messages.consoleMessages.findConfigPlatforms);
+        messages.outputMessage(messages.consoleMessages.findConfigPlatforms);
         var platforms = methods.findConfigPlatforms(config);
         
-        console.log(messages.consoleMessages.installingConfigPlatforms);
+        messages.outputMessage(messages.consoleMessages.installingConfigPlatforms);
         methods.installPlatforms(platforms, options);
     }
 };
 
 exports.sync = function(config, options) {
     options.noForce = true;
-    options.soft = true; // TODO: Implement version checking for hard sync
     
-    console.log(messages.consoleMessages.findingInstalled);
+    messages.outputMessage(messages.consoleMessages.findingInstalled);
     var installedPlugins = methods.findInstalledPlugins();
     
-    console.log(messages.consoleMessages.findingConfigPlugins);
+    messages.outputMessage(messages.consoleMessages.findingConfigPlugins);
     var configPlugins = methods.findConfigPlugins(config);
     
-    console.log(messages.consoleMessages.findInstalledPlatforms);
+    messages.outputMessage(messages.consoleMessages.findInstalledPlatforms);
     var installedPlatforms = methods.findInstalledPlatforms();
     
-    console.log(messages.consoleMessages.findConfigPlatforms);
+    messages.outputMessage(messages.consoleMessages.findConfigPlatforms);
     var configPlatforms = methods.findConfigPlatforms(config);
     var i;
     
@@ -73,74 +72,74 @@ exports.sync = function(config, options) {
     // Remove console messages
     if (!options.noRemove) {
         if (resultPlugins.remove.length > 0) {
-            console.log(messages.consoleMessages.diffPluginsToRemove);
+            messages.outputMessage(messages.consoleMessages.diffPluginsToRemove);
             for (i = 0; i < resultPlugins.remove.length; i++) {
                 console.log(resultPlugins.remove[i].name);
             }
         } else {
-            console.log(messages.consoleMessages.noPluginsToRemove);
+            messages.outputMessage(messages.consoleMessages.noPluginsToRemove);
         }
         
         if (resultPlatforms.remove.length > 0) {
-            console.log(messages.consoleMessages.diffPlatformsToRemove);
+            messages.outputMessage(messages.consoleMessages.diffPlatformsToRemove);
             for (i = 0; i < resultPlatforms.remove.length; i++) {
                 console.log(resultPlatforms.remove[i].name);
             }
         } else {
-            console.log(messages.consoleMessages.noPlatformsToRemove);
+            messages.outputMessage(messages.consoleMessages.noPlatformsToRemove);
         }
     }
     
     // Add console messages
     if (!options.noAdd) {
         if (resultPlugins.install.length > 0) {
-            console.log(messages.consoleMessages.diffPluginsToAdd);
+            messages.outputMessage(messages.consoleMessages.diffPluginsToAdd);
             for (i = 0; i < resultPlugins.install.length; i++) {
                 console.log(resultPlugins.install[i].name);
             }
         } else {
-            console.log(messages.consoleMessages.noPluginsToAdd);
+            messages.outputMessage(messages.consoleMessages.noPluginsToAdd);
         }
         
         if (resultPlatforms.install.length > 0) {
-            console.log(messages.consoleMessages.diffPlatformsToAdd);
+            messages.outputMessage(messages.consoleMessages.diffPlatformsToAdd);
             for (i = 0; i < resultPlatforms.install.length; i++) {
                 console.log(resultPlatforms.install[i].name);
             }
         } else {
-            console.log(messages.consoleMessages.noPlatformsToAdd);
+            messages.outputMessage(messages.consoleMessages.noPlatformsToAdd);
         }
     }
     
     // Remove platforms
     if (!options.noRemove && resultPlatforms.remove.length > 0) {
-        console.log(messages.consoleMessages.removingInstalledPlatforms);
+        messages.outputMessage(messages.consoleMessages.removingInstalledPlatforms);
         methods.removePlatforms(resultPlatforms.remove, options);
     } else if (resultPlatforms.remove.length > 0) {
-        console.log(messages.consoleMessages.skipRemove);
+        messages.outputMessage(messages.consoleMessages.skipRemove);
     }
 
     // Remove plugins
     if (!options.noRemove && resultPlugins.remove.length > 0) {
-        console.log(messages.consoleMessages.removingInstalled);
+        messages.outputMessage(messages.consoleMessages.removingInstalled);
         methods.removePlugins(resultPlugins.remove, options);
     } else if (resultPlugins.remove.length > 0){
-        console.log(messages.consoleMessages.skipRemove);
+        messages.outputMessage(messages.consoleMessages.skipRemove);
     }
     
     // Add plugins
     if (!options.noAdd && resultPlugins.install.length > 0) {
-        console.log(messages.consoleMessages.installingConfigPlugins);
+        messages.outputMessage(messages.consoleMessages.installingConfigPlugins);
         methods.installPlugins(resultPlugins.install, options);
     } else if(resultPlugins.install.length > 0) {
-        console.log(messages.consoleMessages.skipAdd);
+        messages.outputMessage(messages.consoleMessages.skipAdd);
     }
     
     // Add platforms
     if (!options.noAdd && resultPlatforms.install.length > 0) {
-        console.log(messages.consoleMessages.installingConfigPlatforms);
+        messages.outputMessage(messages.consoleMessages.installingConfigPlatforms);
         methods.installPlatforms(resultPlatforms.install, options);
     } else if (resultPlatforms.install.length > 0) {
-        console.log(messages.consoleMessages.skipAdd);
+        messages.outputMessage(messages.consoleMessages.skipAdd);
     }
 };